+++ frontend/src/views/components/post/Post.css
@@ -52,23 +52,27 @@
 
 .content,
 .contentLink {
-  position: absolute;
+  position: relative;
   border: 0;
   outline: 0;
   border-radius: 20px;
 }
 
 .content {
-  top: 60px;
-  left: 20px;
+  box-sizing: border-box;
+  top: 0px;
+  left: 50%;
+  margin-left: -150px;
   width: 300px;
   height: 300px;
+  padding-top: 0.1px;
   box-shadow: 0 0 15px 0 rgba(80, 80, 80, 30%);
   text-align: center;
   font-weight: bold;
 }
 
 .contentLink {
+  position: absolute;
   left: 0px;
   width: 100%;
   height: 100%;
