+++ frontend/src/views/modules/common/fakeServer.js
@@ -15,7 +15,7 @@ function MakeDataGenerator(items) {
 
 export const getNextSAMroad = MakeDataGenerator(samroads);
 export const getPostById = postId =>
-  samroads.filter(samroad => samroad.postId === postId)[0];
+  samroads.filter(samroad => samroad.postId === parseInt(postId))[0];
 
 export function signInReq(userId, password) {
   const user = users[userId];
