+++ frontend/src/views/pages/Pages.css
@@ -5,15 +5,3 @@
   color: #202326;
   background: #f0f3f6;
 }
-
-.HeadBack {
-  box-sizing: border-box;
-  z-index: 9999;
-  position: fixed;
-  top: 0;
-  width: 100%;
-  height: 65px;
-  padding-left: 5%;
-  background: #fdfeff;
-  box-shadow: 0px 5px 20px 0px rgba(80, 80, 80, 10%);
-}
