+++ frontend/src/views/pages/chatPages/index.js
@@ -1,0 +1,4 @@
+import ChatsPage from "./ChatsPage";
+import ChattingPage from "./ChattingPage";
+
+export { ChatsPage, ChattingPage };
