+++ frontend/src/views/pages/postPages/PostPage.css
@@ -1,0 +1,25 @@
+.postHeadBack {
+  box-sizing: border-box;
+  z-index: 9999;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  height: 65px;
+  padding-left: 5%;
+  background: #fdfeff;
+  box-shadow: 0px 5px 20px 0px rgba(80, 80, 80, 10%);
+}
+
+.postingInfo {
+  text-align: center;
+  width: 100%;
+  height: 300px;
+}
+
+.postingBack {
+  display: inline-block;
+  padding-top: 100px;
+  width: 80%;
+  height: 100vh;
+  background: #fdfeff;
+}
