+++ frontend/src/views/pages/postPages/WritePage.js
