+++ frontend/src/views/pages/postPages/index.js
@@ -1,0 +1,5 @@
+import PostPage from "./PostPage";
+import WritePage from "./WritePage";
+import "./PostPage.css";
+
+export { PostPage, WritePage };
